Fix: remove g++ warnings

+++ src/pdf/DfCD.cpp
@@ -292,7 +292,7 @@ TlSymmetricMatrix DfCD::getCholeskyVector(const TlVector& L_col,
                                           const PQ_PairArray& I2PQ)
 {
     const index_type numOfItilde = L_col.getSize();
-    assert(numOfItilde == I2PQ.size());
+    assert(static_cast<std::size_t>(numOfItilde) == I2PQ.size());
 
     TlSymmetricMatrix answer(this->m_nNumOfAOs);
     for (index_type i = 0; i < numOfItilde; ++i) {
